Address a limitation in AIR build system.
As per request of Laura Gouveia (mail discussions on 22/10/2019)

AIR doesn't check during 'make' for a changed target board; it
depends on a 'make clean' done first, if the chosen configuration changes.

To make the final delivery use LEON4-specific functionality, it is
therefore not sufficient to choose a different number in the second
question during configuration and just commit an updated config;
we instead need to somehow trigger this 'make clean' - BUT not always!
We only want to do this once, otherwise every 'Update-TASTE.sh' will
take a lot longer, because AIR would be rebuilt from scratch every time.

Solution: check if there is a previously configured AIR; and if so,
whether it was set for leon4. If it wasn't, 'make clean' first.